requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.6
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
blinker==1.6.3
    # via flask
charset-normalizer==3.3.2
    # via aiohttp
click==8.1.7
    # via flask
colorlog==6.7.0
    # via -r requirements.in
discord-argparse==1.0.1
    # via -r requirements.in
discord-pretty-help==2.0.7
    # via -r requirements.in
discord-py==2.3.2
    # via
    #   -r requirements.in
    #   discord-argparse
flask==3.0.0
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via yarl
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
sqlitedict==2.1.0
    # via -r requirements.in
toml==0.10.2
    # via -r requirements.in
werkzeug==3.0.1
    # via flask
yarl==1.9.2
    # via aiohttp
